+++ tests/Feature/Framework/CustomCacheDirectory/FeatureTest.php
@@ -1,6 +1,6 @@
 <?php
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace GacelaTest\Feature\Framework\CustomCacheDirectory;
 
@@ -8,6 +8,7 @@ use Gacela\Framework\Bootstrap\GacelaConfig;
 use Gacela\Framework\ClassResolver\ClassNameCache;
 use Gacela\Framework\ClassResolver\DocBlockService\CustomServicesCache;
 use Gacela\Framework\Gacela;
+use GacelaTest\Feature\Util\DirectoryUtil;
 use PHPUnit\Framework\TestCase;
 
 final class FeatureTest extends TestCase
@@ -16,7 +17,13 @@ final class FeatureTest extends TestCase
     {
         Gacela::bootstrap(__DIR__, static function (GacelaConfig $config): void {
             $config->addAppConfig('config/*.php');
+            $config->setCacheDirectory('custom-caching-dir');
         });
+    }
+
+    public function tearDown(): void
+    {
+        DirectoryUtil::removeDir(__DIR__ . '/custom-caching-dir');
     }
 
     public function test_custom_caching_dir(): void